# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Araponga API",
    "version": "v1",
    "description": "Araponga é uma plataforma comunitária orientada a território. Esta API expõe recursos de território, feed comunitário, entidades do mapa e saúde do território. O objetivo do MVP é viabilizar cadastro, visualização e curadoria comunitária com segurança e governança mínima.",
    "contact": {
      "name": "Araponga (maintainers)"
    },
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/api/v1/listings": {
      "get": {
        "tags": [
          "Listings"
        ],
        "summary": "Busca listings no território.",
        "parameters": [
          {
            "name": "territoryId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "22222222-2222-2222-2222-222222222222"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "PRODUCT"
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "mel"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Listings"
        ],
        "summary": "Cria um listing (produto ou serviço).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingRequest"
              },
              "examples": {
                "produto": {
                  "summary": "Listing de produto",
                  "value": {
                    "territoryId": "22222222-2222-2222-2222-222222222222",
                    "storeId": "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
                    "type": "PRODUCT",
                    "title": "Mel artesanal 500g",
                    "description": "Mel produzido localmente.",
                    "category": "Alimentos",
                    "tags": "mel,organico",
                    "pricingType": "FIXED",
                    "priceAmount": 32.5,
                    "currency": "BRL",
                    "unit": "unidade",
                    "status": "ACTIVE"
                  }
                },
                "servico": {
                  "summary": "Listing de serviço",
                  "value": {
                    "territoryId": "22222222-2222-2222-2222-222222222222",
                    "storeId": "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
                    "type": "SERVICE",
                    "title": "Aula de surf",
                    "description": "Aula individual de 1h.",
                    "category": "Esportes",
                    "tags": "surf,aula",
                    "pricingType": "HOURLY",
                    "priceAmount": 80,
                    "currency": "BRL",
                    "unit": "hora",
                    "status": "ACTIVE"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/listings/{id}": {
      "get": {
        "tags": [
          "Listings"
        ],
        "summary": "Detalhe de um listing.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Listings"
        ],
        "summary": "Atualiza um listing.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/listings/{id}/archive": {
      "post": {
        "tags": [
          "Listings"
        ],
        "summary": "Arquiva um listing.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateListingRequest": {
        "required": [
          "territoryId",
          "storeId",
          "type",
          "title",
          "pricingType"
        ],
        "type": "object",
        "properties": {
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "pricingType": {
            "type": "string"
          },
          "priceAmount": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateListingRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "pricingType": {
            "type": "string",
            "nullable": true
          },
          "priceAmount": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ListingResponse": {
        "required": [
          "id",
          "territoryId",
          "storeId",
          "type",
          "title",
          "pricingType",
          "status",
          "createdAtUtc",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "pricingType": {
            "type": "string"
          },
          "priceAmount": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}